refactor : BaseballNumberGenerator는 Baseball 도메인 영역에 속한 유틸이라서 domain영역으로 이동
