+++ src/main/java/domain/BaseballNumberGenerator.java
@@ -1,4 +1,4 @@
-package util;
+package domain;
 
 import java.util.Arrays;
 import java.util.Collections;
