+++ src/test/java/util/BaseballNumberGeneratorTest.java
@@ -1,5 +1,6 @@
 package util;
 
+import domain.BaseballNumberGenerator;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
